requirements.txt
Flask
google-generativeai
python-dotenv
flask-cors
flask-limiter
gunicorn
